Reduce duplication in test fixtures
